a stab at making file writes safer… (#301)
…and maybe there are better ways using NO_FOLLOW.

it's definitely something that wants more testing.
