+++ git-worktree/Cargo.toml
@@ -30,6 +30,9 @@ bstr = { version = "0.2.13", default-features = false }
 document-features = { version = "0.2.0", optional = true }
 symlink = "0.1.0"
 
+[target.'cfg(unix)'.dependencies]
+libc = "0.2.119"
+
 [dev-dependencies]
 git-testtools = { path = "../tests/tools" }
 git-odb = { path = "../git-odb" }
